refactor: rename shared to empty-connection
