+++ packages/graphql/src/schema/resolvers/query/empty-connection.ts
